+++ scripts/screens/ChooseAdoptiveParentScreen.py
@@ -7,9 +7,7 @@ import pygame_gui.elements
 from scripts.cat.cats import Cat
 from scripts.game_structure import image_cache
 from scripts.game_structure.propagating_thread import PropagatingThread
-from scripts.game_structure.ui_elements import (
-    UISpriteButton,
-)
+from ..ui.elements.sprite_button import UISpriteButton
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
