+++ scripts/screens/LeaderDenScreen.py
@@ -15,9 +15,7 @@ from scripts.clan_package.settings.clan_settings import (
 )
 from scripts.game_structure import constants
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import (
-    UISpriteButton,
-)
+from scripts.ui.elements.sprite_button import UISpriteButton
 from scripts.ui.elements.image_button import UIImageButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.Screens import Screens
