+++ scripts/screens/MedDenScreen.py
@@ -8,10 +8,10 @@ from scripts.cat.cats import Cat
 from scripts.clan_resources.herb.herb_supply import MESSAGES
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
-    UISpriteButton,
     UITextBoxTweaked,
     UIModifiedImage,
 )
+from ..ui.elements.sprite_button import UISpriteButton
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
