+++ scripts/screens/MediationScreen.py
@@ -8,9 +8,9 @@ import pygame_gui.elements
 from scripts.cat.cats import Cat
 from scripts.game_structure import image_cache, game
 from scripts.game_structure.ui_elements import (
-    UISpriteButton,
     UIRelationDisplay,
 )
+from ..ui.elements.sprite_button import UISpriteButton
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
