+++ scripts/screens/PatrolScreen.py
@@ -8,9 +8,7 @@ import pygame_gui
 from scripts.cat.cats import Cat
 from scripts.events_module.patrol.patrol import Patrol
 from scripts.game_structure import game
-from scripts.game_structure.ui_elements import (
-    UISpriteButton,
-)
+from ..ui.elements.sprite_button import UISpriteButton
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
